+++ components/atoms/buttons/text/primary.tsx
@@ -1,9 +1,10 @@
 'use client'
 
-import { useTheme, alpha } from '@mui/material/styles'
+import { alpha } from '@mui/material/styles'
 import { Button, Typography } from '@mui/material'
 
 // Types
+import type { Theme } from '@mui/material'
 import type { TextButtonProps } from '@/types/components/atoms/buttons/text'
 
 // Helpers
@@ -15,7 +16,6 @@ const PrimaryButtonAtom = (props: TextButtonProps) => {
 
 	// Varaibles
 	const { t } = useTranslation(lng)
-	const theme = useTheme()
 
 	return (
 		<Button
@@ -25,28 +25,38 @@ const PrimaryButtonAtom = (props: TextButtonProps) => {
 			startIcon={startIcon}
 			endIcon={endIcon}
 			onClick={onClick}
-			sx={{
+			sx={(theme: Theme) => ({
 				py: 1.25,
 				px: 2.5,
 				fontSize: 16,
 				textAlign: 'inherit',
 				textTransform: 'inherit',
-				color: theme.palette.mode === 'dark' ? color[100] : color[700],
+				color: color[700],
 				bgcolor: alpha(color[500], 0.15),
 				borderRadius: 1,
 				boxShadow: `${color[500]} 0 -3px 0 inset`,
 				transition: 'all .15s ease',
+				...theme.applyStyles('dark', {
+					color: color[100]
+				}),
 				'&:hover': {
-					color: theme.palette.mode === 'dark' ? color[50] : color[500],
-					bgcolor: alpha(color[500], theme.palette.mode === 'dark' ? 0.2 : 0.1),
-					boxShadow: `${color[300]} 0 -3px 0 inset`
+					color: color[500],
+					bgcolor: alpha(color[500], 0.1),
+					boxShadow: `${color[300]} 0 -3px 0 inset`,
+					...theme.applyStyles('dark', {
+						color: color[50],
+						bgcolor: alpha(color[500], 0.2)
+					})
 				},
 				'&:active': {
-					color: theme.palette.mode === 'dark' ? color[200] : color[700],
+					color: color[700],
 					boxShadow: `${color[700]} 0 0 0 1px inset, ${color[700]} 0 -3px 0 inset`,
-					transform: 'translateY(2px)'
+					transform: 'translateY(2px)',
+					...theme.applyStyles('dark', {
+						color: color[200]
+					})
 				}
-			}}
+			})}
 		>
 			<Typography component="span" fontSize={16} fontWeight={600} flexGrow={1}>
 				{t(title)}