+++ components/organisms/appbar.tsx
@@ -3,11 +3,12 @@
 import { useState } from 'react'
 import dynamic from 'next/dynamic'
 import { useTheme, alpha } from '@mui/material/styles'
-import { useMediaQuery, AppBar, Container, Toolbar, Grid } from '@mui/material'
+import { useMediaQuery, AppBar, Container, Toolbar, Grid2 as Grid } from '@mui/material'
 import { blueGrey } from '@mui/material/colors'
 import { Menu as MenuIcon } from '@mui/icons-material'
 
 // Types
+import type { Theme } from '@mui/material/styles'
 import type { AppbarProps } from '@/types/components/organisms/appbar'
 
 // Components
@@ -23,23 +24,25 @@ const AppbarOrganism = (props: AppbarProps) => {
 	const { lng } = props
 
 	// Variables
-	const theme = useTheme()
+	const muiTheme = useTheme()
 	const [open, setOpen] = useState<boolean>(false)
 	const [menuParent, setMenuParent] = useState<string>('')
-	const greaterThanMedium = useMediaQuery(theme.breakpoints.up('md'))
-	useMediaQuery('(prefers-color-scheme: dark)')
+	const greaterThanMedium = useMediaQuery(muiTheme.breakpoints.up('md'))
 
 	return (
 		<HideOnScroll onlyDesktop={true}>
 			<AppBar
 				component="nav"
-				sx={{
+				sx={(theme: Theme) => ({
 					color: 'inherit',
 					backgroundImage: 'none',
-					bgcolor: alpha(theme.palette.background.default, 0.17),
+					bgcolor: alpha('#fafafa', 0.17),
 					backdropFilter: 'blur(20px)',
-					boxShadow: 'none'
-				}}
+					boxShadow: 'none',
+					...theme.applyStyles('dark', {
+						bgcolor: alpha('#131313', 0.17)
+					})
+				})}
 			>
 				<Container maxWidth="xl">
 					<Toolbar
@@ -52,7 +55,7 @@ const AppbarOrganism = (props: AppbarProps) => {
 
 						<Grid container justifyContent={{ xs: 'end', md: 'space-between' }} flexGrow={1} ml={{ xs: 1, md: 5 }}>
 							{!greaterThanMedium && (
-								<Grid item>
+								<Grid>
 									<IconButtonAtom color={blueGrey} icon={<MenuIcon />} onClick={() => setOpen(!open)} />
 
 									<DrawerMolecule lng={lng} open={open} setOpen={setOpen} menuParent={menuParent} setMenuParent={setMenuParent}>